for graphs

+++ DataSkeletonizationNew/include/TestModule/Statistics/SimpleStatistic.h
@@ -41,7 +41,7 @@ public:
     }
     double returnAvg()
     {
-    return success/caseNumber;
+    return (double) success/ (double) caseNumber;
     }
